[Chore] #240 WriteStudyDao로 이름 변경
